+++ changelog.md
@@ -1,4 +1,4 @@
-
+### v2.2.0
 ---
 * **Tier 19**
  * Added support for Nighthold Tier 19 tokens.